Change default for shakeFiles to ".shake" instead of "."
With the current default of "." the database was created as "..database".
Since the code for writing the database and filenames mentions a filename
as the base (rather than a directory) it's probably best to use a filename
as the default for shakeFiles.  So by default Shake writes the database to
.shake.database and the journal to .shake.database which is nicely
descriptive.