Better Shake reports This patch adds support for building HTML profiling reports directly from within Shake. Now you can always generate a resulting report by setting 'ShakeOptions.shakeDump' to 'Just filepath' to output the report to 'filepath'.
The output is completely standalone. It is pretty thanks to Bootstrap,
jQuery and flot. It even works on mobile devices (or at least one of them!)

This patch adds a dependency on HStringTemplate. The reporting module
also makes use of the -XCPP extension, to detect when we're building
with cabal.