Calculate correct relative paths from fsatrace output.
With fsatrace linting, a command that writes a file, unknown to
the build system, should then cause a linting error if that file
is later `need`ed. Previously, shake would fail to give this
expected linting error when the file in question if not under
the current working directory.

The reason for this is fsatrace gives absolute file paths and
`makeRelative` will often fail to convert to a relative path e.g:

    > makeRelative "/root/a/" "/root/b/file.out"
    "/root/b/file.out"

This is because `makeRelative` is pure and cannot guarantee a
correct relative path incase "a" is a symlink (i.e.
"/root/a/../b/file.out" is not necessarily equal to
"/root/b/file.out"). This commit fixes the bug by introducing the
function `makeRelativeEx` that guarantees a relative path is found
when one exists (at the cost of being impure).

    > -- Given that "/root/a/" is not a symlink
    > makeRelativeEx "/root/a/" "/root/b/file.out"
    Just "../b/file.out"

    > -- Given that "/root/c/" is a symlink to "/root/e/f/g/"
    > makeRelativeEx "/root/c/" "/root/b/file.out"
    Just "../../../b/file.out"