shake: embed data files into 'shake' executable
This patch directly embeds the set of needed re-distributable artifacts
in the Shake repository into the 'shake' executable by default, and,
optionally, the 'shake' library as well.

Without this change, Shake binaries cannot build HTML reports without
the Haskell toolchain installed. This means that users of Shake for e.g.
ninja compatibility are unable to build reports transparently.

There have been ideas floated around to embed data files at the end of
the binary, but this seems needlessly complicated. Instead, this patch
simply embeds the files in the binary at compile time if asked, using
Template Haskell.

One complicating factor that made this change hard in the past is that
GHC is moving to use Hadrian (and Shake) by default, and so adding a
stage2 feature dependency (TemplateHaskell requires GHCi) seriously
hurts the ability to bootstrap GHC: you would need an entire stage2
cross toolchain to build Hadrian, for instance.

However, there are some crucial distinctions to make here, and some
factors that make this change not so bad:

  - The `shake` executable, which provides Ninja compatibility, is NOT
  bound by the Hadrian/Haskell stage2 requirements. This means that even
  e.g.  CMake or external users of Shake can still benefit from such a
  change, if it is done carefully.

  - The `shake` library, which is what developers use to write build
  systems, MAY be bound by the stage2 restriction, i.e. cannot use
  Template Haskell, due to platform restrictions. But it also may not
  be; a user on x86_64-linux who is writing a custom build system may
  well want to deliver it to their developers, also running
  x86_64-linux.

  - Users of Haskell tooling, such as Stack or Cabal (new-build),
  generally have the ability to set per-package compilation flags for
  their projects. For instance, a single project using `cabal new-build`
  can globally enforce the usage of Shake, but only compiled with a
  specific set of build-time flags. While this wasn't always available,
  it is now generally widely available.

The third mitigating factor is what motivates this solution as the
simplest and easiest: to simply add support and help users in group 1
(currently myself), and also some of the users in group 2 by simply
requiring them to set flags for the dependent package.

In this setup, the `shake` library, by default, still DOES NOT use file
embedding. Rather, setting the flag `-fembed-files` at build time will
enable support for it in the library. Users of Shake who want to create
binaries that have profiling support "included" can simply enforce their
projects to have this flag enabled, and their binaries will always
support this feature.

The `shake` executable, however, ALWAYS has support for file-embedding
built in. This executable is generally useful only for Ninja
compatibility (or general demo purposes), so adding this requirement
does not seem especially onerous, and improves the lives of any user who
is currently depends on it.

This patch has been tested on a full CMake/Ninja build on a standalone
Ubuntu 16.04 machine, and was able to generate a working HTML build
report.

Closes #268. Closes #413.

Signed-off-by: Austin Seipp <as@fastly.com>